clean up + docs
